+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 1.2.10
+
+ * Update levels for Hikes and Nights away staged badges (released April 2014)
+ * Add levels for new Time on the water staged badge (released April 2014)
+
 ## Version 1.2.9
 
  * Add osm\_long\_key to badges to make getting badge stock easier